+++ client/.eslintcache
@@ -1,0 +1,1 @@
+[{"C:\\Users\\User\\Documents\\GitHub\\M7011E_project\\client\\src\\index.js":"1","C:\\Users\\User\\Documents\\GitHub\\M7011E_project\\client\\src\\reportWebVitals.js":"2","C:\\Users\\User\\Documents\\GitHub\\M7011E_project\\client\\src\\App.js":"3","C:\\Users\\User\\Documents\\GitHub\\M7011E_project\\client\\src\\pages\\Start.js":"4","C:\\Users\\User\\Documents\\GitHub\\M7011E_project\\client\\src\\pages\\Register.js":"5","C:\\Users\\User\\Documents\\GitHub\\M7011E_project\\client\\src\\pages\\Login.js":"6","C:\\Users\\User\\Documents\\GitHub\\M7011E_project\\client\\src\\components\\Navigation.js":"7","C:\\Users\\User\\Documents\\GitHub\\M7011E_project\\client\\src\\components\\LoginForm.js":"8","C:\\Users\\User\\Documents\\GitHub\\M7011E_project\\client\\src\\components\\register\\RegisterSignup.js":"9","C:\\Users\\User\\Documents\\GitHub\\M7011E_project\\client\\src\\components\\register\\RegisterSuccess.js":"10","C:\\Users\\User\\Documents\\GitHub\\M7011E_project\\client\\src\\components\\register\\validateRegistration.js":"11","C:\\Users\\User\\Documents\\GitHub\\M7011E_project\\client\\src\\components\\register\\useRegister.js":"12"},{"size":540,"mtime":1610466315023,"results":"13","hashOfConfig":"14"},{"size":375,"mtime":1607942678024,"results":"15","hashOfConfig":"14"},{"size":737,"mtime":1610466315010,"results":"16","hashOfConfig":"14"},{"size":171,"mtime":1610466315040,"results":"17","hashOfConfig":"14"},{"size":572,"mtime":1610466551550,"results":"18","hashOfConfig":"14"},{"size":215,"mtime":1610477422503,"results":"19","hashOfConfig":"14"},{"size":418,"mtime":1610475847040,"results":"20","hashOfConfig":"14"},{"size":2051,"mtime":1610477404287,"results":"21","hashOfConfig":"14"},{"size":3570,"mtime":1610477615809,"results":"22","hashOfConfig":"14"},{"size":498,"mtime":1610468009068,"results":"23","hashOfConfig":"14"},{"size":793,"mtime":1610466452898,"results":"24","hashOfConfig":"14"},{"size":931,"mtime":1610475541645,"results":"25","hashOfConfig":"14"},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},"1ou20t3",{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"37","messages":"38","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"39","messages":"40","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"41","messages":"42","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"43","messages":"44","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"45","messages":"46","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":"47","usedDeprecatedRules":"28"},{"filePath":"48","messages":"49","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"50","messages":"51","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"C:\\Users\\User\\Documents\\GitHub\\M7011E_project\\client\\src\\index.js",[],["52","53"],"C:\\Users\\User\\Documents\\GitHub\\M7011E_project\\client\\src\\reportWebVitals.js",[],"C:\\Users\\User\\Documents\\GitHub\\M7011E_project\\client\\src\\App.js",[],"C:\\Users\\User\\Documents\\GitHub\\M7011E_project\\client\\src\\pages\\Start.js",[],"C:\\Users\\User\\Documents\\GitHub\\M7011E_project\\client\\src\\pages\\Register.js",[],"C:\\Users\\User\\Documents\\GitHub\\M7011E_project\\client\\src\\pages\\Login.js",[],"C:\\Users\\User\\Documents\\GitHub\\M7011E_project\\client\\src\\components\\Navigation.js",[],"C:\\Users\\User\\Documents\\GitHub\\M7011E_project\\client\\src\\components\\LoginForm.js",[],"C:\\Users\\User\\Documents\\GitHub\\M7011E_project\\client\\src\\components\\register\\RegisterSignup.js",[],"C:\\Users\\User\\Documents\\GitHub\\M7011E_project\\client\\src\\components\\register\\RegisterSuccess.js",["54"],"import React from 'react'\r\n\r\nconst RegisterSuccess = () => {\r\n    return (\r\n        <div className=\"form-how\">\r\n            <div className=\"register-success\">Your registration was successfull </div>\r\n            <img src=\"img/WindTurbine.svg\" width=\"35%\" alt=\"turbine-image\" className=\"form-img-2\" />\r\n            <span className=\"success-input-login\">\r\n                Login in <a href=\"/login\">here</a> \r\n            </span>\r\n        </div>\r\n        \r\n    )\r\n}\r\n\r\nexport default RegisterSuccess\r\n","C:\\Users\\User\\Documents\\GitHub\\M7011E_project\\client\\src\\components\\register\\validateRegistration.js",[],"C:\\Users\\User\\Documents\\GitHub\\M7011E_project\\client\\src\\components\\register\\useRegister.js",["55"],{"ruleId":"56","replacedBy":"57"},{"ruleId":"58","replacedBy":"59"},{"ruleId":"60","severity":1,"message":"61","line":7,"column":13,"nodeType":"62","endLine":7,"endColumn":101},{"ruleId":"63","severity":1,"message":"64","line":34,"column":7,"nodeType":"65","endLine":34,"endColumn":15,"suggestions":"66"},"no-native-reassign",["67"],"no-negated-in-lhs",["68"],"jsx-a11y/img-redundant-alt","Redundant alt attribute. Screen-readers already announce `img` tags as an image. You don’t need to use the words `image`, `photo,` or `picture` (or any specified custom words) in the alt prop.","JSXOpeningElement","react-hooks/exhaustive-deps","React Hook useEffect has missing dependencies: 'callback' and 'submitting'. Either include them or remove the dependency array. If 'callback' changes too often, find the parent component that defines it and wrap that definition in useCallback.","ArrayExpression",["69"],"no-global-assign","no-unsafe-negation",{"desc":"70","fix":"71"},"Update the dependencies array to be: [callback, errors, submitting]",{"range":"72","text":"73"},[825,833],"[callback, errors, submitting]"]
