--- client/src/Register.css
@@ -1,3 +1,0 @@
-.form-inputs p {
-    color: red
-}
