+++ client/src/components/register/Register.css
@@ -1,0 +1,8 @@
+.form-inputs p {
+    color: red
+}
+
+.form-how {
+    margin-left: 35%;
+    width: 30%;
+}
