+++ client/src/components/register/RegisterSuccess.js
@@ -6,7 +6,7 @@ const RegisterSuccess = () => {
             <div className="register-success">Your registration was successfull </div>
             <img src="img/WindTurbine.svg" width="35%" alt="turbine-image" className="form-img-2" />
             <span className="success-input-login">
-                Login in <a href="#">here</a> 
+                Login in <a href="/login">here</a> 
             </span>
         </div>
