+++ client/src/components/register/validateRegistration.js
