+++ client/src/pages/Login.js
@@ -4,7 +4,7 @@ import './Login.css'
 function Login() {
     return (
       <div className="Login">
-        <h1>Login page</h1>
+
         <LoginForm />
       </div>
     );
