+++ client/src/pages/ManagerPage.js
@@ -1,0 +1,12 @@
+
+
+function ManagerPage() {
+    return (
+      <div className="ManagerPage">
+
+        
+      </div>
+    );
+}
+
+export default ManagerPage;
