+++ client/src/pages/ProsumerPage.js
@@ -1,0 +1,12 @@
+
+
+function ProsumerPage() {
+    return (
+      <div className="ProsumerPage">
+
+        
+      </div>
+    );
+}
+
+export default ProsumerPage;
