+++ package.json
@@ -9,6 +9,7 @@
   },
   "dependencies": {
     "bluebird": "^3.7.2",
+    "bootstrap": "^4.5.3",
     "compute-erf": "^3.0.3",
     "compute-erfinv": "^3.0.1",
     "compute-mean": "^3.0.0",
@@ -29,6 +30,8 @@
     "jsonwebtoken": "^8.5.1",
     "morgan": "~1.9.1",
     "node-fetch": "^2.6.1",
+    "react-bootstrap": "^1.4.3",
+    "react-router-dom": "^5.2.0",
     "sqlite3": "^5.0.0"
   }
 }
